fix build errors on MinGW
The following changes have been made:

In /CMakeLists.txt, the rule for finding the `readline` package
has been changed to use a default location in case the automatic
search fails (and on MinGW it is always failing). It includes a
check to pick the correct mingw root based on the the build
architecture. It is an optional dependency.

/external/unbound/CMakeLists.txt has the fix for finding the winpthread
library on MinGW. It cnnot be automatically found on MinGW, so the
Windows-style path needed to be placed in the unbound CMakeLists.txt
file, to be copied by the CMake build process into the local build
folder.

The epee header files include <boost/bind/placeholders.h> and using
declarations _1 _2 _3 _4, because there were errors related to these
constants being undefined.

The cryptonote_core/blockchain.h and .cpp files have the template
removed from get_transaction_blobs because it caused a linking error
against ryod.exe (a template function cannot be called from a different
source file if its body is defined in another source file).